Fixing bad rewrite behavior for older CouchDB versions.
For some reason, some CouchDB versions (tested with v1.1.0) do not
like correctly typed parameters for rewrite rules - like

    {
    "from": "/live",
    "to": "/_list/live/ranked_games",
    "method": "GET",
    "query": {
      "descending": true
    }

instead the value for the `descending` parameter needs to be set as a
string to make it work. This seems to be a confirmed bug for CouchDB,
see https://issues.apache.org/jira/browse/COUCHDB-1193 for more
detail. However, Membase (CouchDB 1.1.1) seems not have this issue,
but this modification does not break the behavior since arguments are
treated as strings in the end.